InvoiceLayoutLM Extractor/data_generation/requirements.txt
brotli==1.2.0
cffi==2.0.0
cssselect2==0.9.0
Faker==40.8.0
fonttools==4.62.0
greenlet==3.3.2
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.1.1
playwright==1.58.0
PyMuPDF==1.27.2
pycparser==3.0
pydyf==0.12.1
pyee==13.0.1
pyphen==0.17.2
tinycss2==1.5.1
tinyhtml5==2.1.0
typing_extensions==4.15.0
tzdata==2025.3
webencodings==0.5.1
zopfli==0.4.1
